refactor: split new API into multiple files
