+++ packages/antithrow/src/new/index.ts
@@ -1,0 +1,6 @@
+export { Err } from "./err.js";
+export { UnwrapError } from "./errors.js";
+export { Ok } from "./ok.js";
+export { Pending } from "./pending.js";
+export { Result } from "./result.js";
+export type { InferErr, InferOk, Settled } from "./types.js";
